% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoralContestation-package.R
\docType{package}
\name{electoralContestation-package}
\alias{electoralContestation}
\alias{electoralContestation-package}
\title{electoralContestation: Replication package to accompany ""}
\description{
This package accompanies ...
}
\keyword{internal}
